cat-emotion-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]
sqlalchemy
pydantic
numpy
scikit-learn
joblib
librosa
tensorflow
tensorflow-hub
torch
torchvision
torchaudio
requests
google-generativeai
python-dotenv
